refactor: poly something idk
