+++ package.json
@@ -20,6 +20,7 @@
     "konva": "^9.3.22",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "react-error-boundary": "^6.0.0",
     "react-i18next": "^15.6.0",
     "react-icons": "^5.5.0",
     "react-konva": "^19.0.7",
