+++ src/components/Sidebar/FontPicker.tsx
@@ -1,6 +1,5 @@
 import { useEffect } from "react";
 import Select from "./Select";
-import type { onStyleChangeFn } from "./Sidebar";
 
 const localFonts = ["Arial", "Times New Roman"];
 
@@ -29,7 +28,7 @@ const FontPicker = ({
   onStyleChange,
   currentFont,
 }: {
-  onStyleChange: onStyleChangeFn;
+  onStyleChange: (property: string, value: unknown) => void;
   currentFont: string | undefined;
 }) => {
   useEffect(() => {
