--- src/consts/symbol.ts
@@ -1,2 +1,0 @@
-export const defaultWidth = 100;
-export const defaultHeight = 100;
