+++ src/hooks/useSelectedSymbols.ts
@@ -1,0 +1,8 @@
+import { useAppStore } from "../store/store";
+
+export default function useSelectedSymbols() {
+  const elements = useAppStore.use.elements();
+  const selectedIds = useAppStore.use.selectedIds();
+
+  return elements.filter(({ id }) => selectedIds.includes(id));
+}
