+++ src/main.tsx
@@ -5,11 +5,14 @@ import App from "./App.tsx";
 import "./i18n"; // Import i18n configuration
 import { I18nextProvider } from "react-i18next";
 import i18n from "./i18n";
+import { ErrorBoundary } from "react-error-boundary";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
     <I18nextProvider i18n={i18n}>
-      <App />
+      <ErrorBoundary fallback={<p>Nie dziala</p>}>
+        <App />
+      </ErrorBoundary>
     </I18nextProvider>
   </StrictMode>,
 );
